# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"JB Online API Documentation","description":"JB Online OpenApi description","license":{"name":"MIT License","url":"https:\/\/github.com\/NJUCSE17\/JB-Online"},"version":"v1"},"paths":{"\/api\/assignments":{"get":{"tags":["Assignment"],"summary":"Get all ongoing assignments.","operationId":"App\\Http\\Controllers\\API\\Forum\\AssignmentController::getAssignments","responses":{"200":{"description":"Successful operation","content":{"application\/json":{"schema":{"properties":{"assignments":{"description":"An array of assignments","type":"array","items":{"property":null,"$ref":"#\/components\/schemas\/AssignmentInfo"}}},"type":"object"}}}}},"security":[{"passport":[]}]}},"\/api\/assignment\/{assignment_id}\/finish":{"post":{"tags":["Assignment"],"summary":"Mark an assignment as finished and update assignments.","operationId":"App\\Http\\Controllers\\API\\Forum\\AssignmentController::finishAssignment","parameters":[{"name":"assignment_id","in":"path","description":"Numeric ID of the assignment.","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Successful operation","content":{"application\/json":{"schema":{"properties":{"assignments":{"description":"An array of assignments","type":"array","items":{"property":null,"$ref":"#\/components\/schemas\/AssignmentInfo"}}},"type":"object"}}}}},"security":[{"passport":[]}]}},"\/api\/assignment\/{assignment_id}\/reset":{"post":{"tags":["Assignment"],"summary":"Mark an assignment as unfinished and update assignments.","operationId":"App\\Http\\Controllers\\API\\Forum\\AssignmentController::resetAssignment","parameters":[{"name":"assignment_id","in":"path","description":"Numeric ID of the assignment.","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Successful operation","content":{"application\/json":{"schema":{"properties":{"assignments":{"description":"An array of assignments","type":"array","items":{"property":null,"$ref":"#\/components\/schemas\/AssignmentInfo"}}},"type":"object"}}}}},"security":[{"passport":[]}]}},"\/api\/heatmap":{"get":{"tags":["Utils"],"summary":"Get the JSON data of heatmap on index page.","operationId":"App\\Http\\Controllers\\API\\HomeController::heatmap","parameters":[{"name":"st","in":"query","description":"Start time (in timestamp format).","required":false,"schema":{"type":"integer"}},{"name":"ed","in":"query","description":"End time (in timestamp format).","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Successful operation","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/Heatmap"}}}}}}},"\/api\/app":{"get":{"tags":["Utils"],"summary":"Get the JSON data of latest mobile app info.","operationId":"App\\Http\\Controllers\\API\\HomeController::app","responses":{"200":{"description":"Successful operation","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/AppInfo"}}}}}}},"\/oauth\/clients":{"get":{"tags":["OAuth2"],"summary":"Get the list of clients of current user.","responses":{"200":{"description":"Successful operation","content":{"application\/json":{"schema":{"type":"array","items":{"property":null,"$ref":"#\/components\/schemas\/ClientInfo"}}}}}},"security":[{"passport":[]}]},"post":{"tags":["OAuth2"],"summary":"Create a new OAuth2 authorization code grant type client.","requestBody":{"content":{"application\/x-www-form-urlencoded":{"schema":{"required":["name","redirect"],"properties":{"name":{"description":"Name of the client.","type":"string"},"redirect":{"description":"Redirect URI of the client.","type":"string"}},"type":"object"}}}},"responses":{"200":{"description":"Successful operation","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/ClientInfo"}}}}},"security":[{"passport":[]}]}},"\/oauth\/clients\/{client-id}":{"put":{"tags":["OAuth2"],"summary":"Update an OAuth2 client.","parameters":[{"name":"client-id","in":"path","description":"Numeric ID of the client.","required":true,"schema":{"type":"integer"}}],"requestBody":{"content":{"application\/x-www-form-urlencoded":{"schema":{"required":["name","redirect"],"properties":{"name":{"description":"Name of the client.","type":"string"},"redirect":{"description":"Redirect URI of the client.","type":"string"}},"type":"object"}}}},"responses":{"200":{"description":"Successful operation","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/ClientInfo"}}}}},"security":[{"passport":[]}]},"delete":{"tags":["OAuth2"],"summary":"Delete an OAuth2 client.","parameters":[{"name":"client-id","in":"path","description":"Numeric ID of the client.","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Successful operation"}},"security":[{"passport":[]}]}},"\/oauth\/authorize":{"get":{"tags":["OAuth2"],"summary":"Compose an OAuth2 authorization request and get authorization code.","parameters":[{"name":"client_id","in":"query","description":"Numeric ID of the client.","required":true,"schema":{"type":"integer"}},{"name":"redirect_uri","in":"query","description":"Redirect URI of the client.","required":true,"schema":{"type":"string"}},{"name":"response_type","in":"query","description":"Response type of the request.","required":true,"schema":{"type":"string","enum":["code"]}},{"name":"scope","in":"query","description":"Scope requested, (there is no scope registered in JB Online).","required":false,"schema":{"type":"string","enum":[""]}}],"responses":{"401":{"description":"Client authorization failure"},"200":{"description":"Authorization page"},"302":{"description":"Succeed and redirect to callback page"}}}},"\/oauth\/token":{"post":{"tags":["OAuth2"],"summary":"Request for an access token with one-time authorization code.","requestBody":{"content":{"application\/x-www-form-urlencoded":{"schema":{"$ref":"#\/components\/schemas\/AccessRequest"}}}},"responses":{"200":{"description":"Successful access request","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/AccessSuccessfulResponse"}}}},"400":{"description":"Unsuccessful access request","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/AccessUnsuccessfulResponse"}}}}}}}},"components":{"schemas":{"AssignmentInfo":{"properties":{"id":{"description":"Numeric ID of the assignment.","type":"integer"},"course_id":{"description":"Numeric ID of the source course of assignment.","type":"integer"},"name":{"description":"Name of the assignment.","type":"string"},"content":{"description":"Content of the assignment.","type":"string"},"due_time":{"description":"Deadline of the assignment.","type":"string"},"issuer":{"description":"Numeric ID of the issuer of the assignment.","type":"integer"},"finished":{"description":"Whether the assignment is finished.","type":"boolean"}},"type":"object"},"Heatmap":{"properties":{"timestamp":{"description":"The number of entries of a timestamp","type":"integer"}},"type":"object"},"AppInfo":{"properties":{"number":{"description":"Version number","type":"integer"},"name":{"description":"Version name","type":"string"},"info":{"description":"Version info","type":"string"},"link":{"description":"App download link","type":"string"}},"type":"object"},"ClientInfo":{"properties":{"id":{"description":"Numeric ID of the client.","type":"integer"},"user_id":{"description":"Numeric ID of the issuer of the client.","type":"integer"},"name":{"description":"Name of the client.","type":"string"},"secret":{"description":"Secret token of the client.","type":"string"},"redirect":{"description":"Redirect URI of the client.","type":"string"},"personal_access_client":{"description":"Whether this is a personal access type client.","type":"boolean"},"password_client":{"description":"Whether this is a password grant type client.","type":"boolean"},"revoked":{"description":"Whether this client is revoked.","type":"boolean"},"created_at":{"description":"Timestamp of the created time of the client.","type":"string"},"updated_at":{"description":"Timestamp of the last updated time of the client.","type":"string"}},"type":"object"},"AccessRequest":{"required":["grant_type","client_id","client_secret"],"properties":{"grant_type":{"description":"OAuth2 Grant Type","type":"string"},"client_id":{"description":"OAuth2 Client ID","type":"string"},"client_secret":{"description":"OAuth2 Client Secret","type":"string"},"scope":{"description":"OAuth2 Scope Requested (leave it empty)","type":"string"},"redirect_uri":{"description":"(Authorization Code Grant Type) OAuth2 Redirect URI","type":"string"},"code":{"description":"(Authorization Code Grant Type) OAuth2 Authorization Code","type":"string"},"username":{"description":"(Password Grant Type) Username","type":"string"},"password":{"description":"(Password Grant Type) Password","type":"string"}},"type":"object"},"AccessSuccessfulResponse":{"properties":{"access_token":{"description":"Access token.","type":"string"},"token_type":{"description":"Type of the token.","type":"string"},"expires_in":{"description":"Time before the token expires.","type":"string"},"refresh_token":{"description":"Refresh token.","type":"string"},"scope":{"description":"Scopes of the token.","type":"string"}},"type":"object"},"AccessUnsuccessfulResponse":{"properties":{"error":{"description":"Error type.","type":"string"},"message":{"description":"Error message.","type":"string"},"hint":{"description":"Hint to solve the error.","type":"string"}},"type":"object"}},"securitySchemes":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","in":"header","scheme":"https","flows":{"authorizationCode":{"authorizationUrl":"http:\/\/localhost:8000\/oauth\/authorize","tokenUrl":"http:\/\/localhost:8000\/oauth\/token","scopes":[]}}}}}}